+++ src/formatters/AmazonAlexaSsmlFormatter.ts
@@ -24,26 +24,43 @@ export class AmazonAlexaSsmlFormatter extends SsmlFormatterBase {
       }
       case 'paragraph': {
         if (this.options.includeParagraphTag) {
-          return this.addTag('p', ast.children, true, true, null, lines);
+          return this.addTag('p', ast.children, true, false, null, lines);
         } else {
           this.processAst(ast.children, lines);
           return lines;
         }
       }
       case 'shortBreak': {
-        return this.addBreak(ast, lines);
+        const time = ast.children[0].allText;
+        return this.addBreakTime(lines, time);
       }
       case 'shortEmphasisModerate': {
-        return this.addEmphasis(ast, lines, 'moderate');
+        const text = ast.children[0].allText;
+        return this.addEmphasis(lines, text, 'moderate');
       }
       case 'shortEmphasisStrong': {
-        return this.addEmphasis(ast, lines, 'strong');
+        const text = ast.children[0].allText;
+        return this.addEmphasis(lines, text, 'strong');
       }
       case 'shortEmphasisNone': {
-        return this.addEmphasis(ast, lines, 'none');
+        const text = ast.children[0].allText;
+        return this.addEmphasis(lines, text, 'none');
       }
       case 'shortEmphasisReduced': {
-        return this.addEmphasis(ast, lines, 'reduced');
+        const text = ast.children[0].allText;
+        return this.addEmphasis(lines, text, 'reduced');
+      }
+      case 'textModifier': {
+        const text = ast.children[0].allText;
+        const key = ast.children[1].allText;
+        const value = ast.children.length === 3 ? ast.children[2].allText : '';
+
+        if (key === 'emphasis') {
+          const level = value || 'moderate';
+          return this.addEmphasis(lines, text, level);
+        }
+
+        return lines;
       }
       case 'simpleLine': {
         this.processAst(ast.children, lines);
@@ -62,10 +79,6 @@ export class AmazonAlexaSsmlFormatter extends SsmlFormatterBase {
       }
       case 'plainText': {
         lines.push(ast.allText);
-        return lines;
-      }
-      case 'annotations:': {
-        // lines.push(ast.allText);
         return lines;
       }