+++ tests/emphasis-standard.spec.ts
@@ -1,0 +1,227 @@
+// tslint:disable-next-line:import-name
+import dedent from 'ts-dedent';
+import { SpeechMarkdown } from '../src/SpeechMarkdown';
+
+describe('emphasis-standard-moderate', () => {
+
+  const speech = new SpeechMarkdown();
+
+  const markdown = dedent`
+    A (moderate)[emphasis] level
+    A (moderate)[emphasis:"moderate"] level
+  `;
+
+  test('converts to SSML - Amazon Alexa', () => {
+
+    const options = {
+      platform: 'amazon-alexa'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      A <emphasis level="moderate">moderate</emphasis> level
+      A <emphasis level="moderate">moderate</emphasis> level
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to SSML - Google Assistant', () => {
+
+    const options = {
+      platform: 'google-assistant'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      A <emphasis level="moderate">moderate</emphasis> level
+      A <emphasis level="moderate">moderate</emphasis> level
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to Plain Text', () => {
+
+    const options = {
+    };
+    const text = speech.toText(markdown, options);
+
+    const expected = dedent`
+      A moderate level
+      A moderate level
+    `;
+
+    expect(text).toBe(expected);
+  });
+
+});
+
+describe('emphasis-standard-strong', () => {
+
+  const speech = new SpeechMarkdown();
+
+  const markdown = dedent`
+    A (strong)[emphasis:"strong"] level
+  `;
+
+  test('converts to SSML - Amazon Alexa', () => {
+
+    const options = {
+      platform: 'amazon-alexa'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      A <emphasis level="strong">strong</emphasis> level
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to SSML - Google Assistant', () => {
+
+    const options = {
+      platform: 'google-assistant'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      A <emphasis level="strong">strong</emphasis> level
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to Plain Text', () => {
+
+    const options = {
+    };
+    const text = speech.toText(markdown, options);
+
+    const expected = dedent`
+      A strong level
+    `;
+
+    expect(text).toBe(expected);
+  });
+
+});
+
+describe('emphasis-standard-none', () => {
+
+  const speech = new SpeechMarkdown();
+
+  const markdown = dedent`
+    A (none)[emphasis:"none"] level
+  `;
+
+  test('converts to SSML - Amazon Alexa', () => {
+
+    const options = {
+      platform: 'amazon-alexa'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      A <emphasis level="none">none</emphasis> level
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to SSML - Google Assistant', () => {
+
+    const options = {
+      platform: 'google-assistant'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      A <emphasis level="none">none</emphasis> level
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to Plain Text', () => {
+
+    const options = {
+    };
+    const text = speech.toText(markdown, options);
+
+    const expected = dedent`
+      A none level
+    `;
+
+    expect(text).toBe(expected);
+  });
+
+});
+
+describe('emphasis-standard-reduced', () => {
+
+  const speech = new SpeechMarkdown();
+
+  const markdown = dedent`
+    A (reduced)[emphasis:"reduced"] level
+  `;
+
+  test('converts to SSML - Amazon Alexa', () => {
+
+    const options = {
+      platform: 'amazon-alexa'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      A <emphasis level="reduced">reduced</emphasis> level
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to SSML - Google Assistant', () => {
+
+    const options = {
+      platform: 'google-assistant'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      A <emphasis level="reduced">reduced</emphasis> level
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to Plain Text', () => {
+
+    const options = {
+    };
+    const text = speech.toText(markdown, options);
+
+    const expected = dedent`
+      A reduced level
+    `;
+
+    expect(text).toBe(expected);
+  });
+
+});