+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to the speechmarkdown-js project will be documented in this file.
 
+## 0.4.0-alpha.0 - (June 30, 2019)
+### Added
+- Update grammar and formatters for emphasis standard format
+- Add tests to increase coverage
+
 ## 0.3.0-alpha.0 - (June 30, 2019)
 ### Added
 - Update grammar and formatters for emphasis short format